requirements.txt
streamlit
python-dotenv
google-genai
sentence-transformers
faiss-cpu
pypdf
numpy
